Migrate functions to dotfinder.py
