+++ cooltools/dotfinder.py
@@ -2,6 +2,9 @@
 Collection of functions needed for dot-calling
 
 """
+from functools import partial, reduce
+import multiprocess as mp
+
 from scipy.linalg import toeplitz
 from scipy.ndimage import convolve
 from scipy.stats import poisson
@@ -9,6 +12,22 @@ from scipy.sparse import coo_matrix
 import numpy as np
 import pandas as pd
 from sklearn.cluster import Birch
+import cooler
+
+from .lib.numutils import LazyToeplitz, get_kernel
+
+
+# these are the constants from HiCCUPS, that dictate how initial histogramms
+# for lambda-chunking are computed: W1 is the # of logspaced lambda bins,
+# and W2 is maximum "allowed" raw number of contacts per Hi-C heatmap bin:
+HiCCUPS_W1_MAX_INDX = 40
+
+# HFF combined exceeded this limit ...
+HiCCUPS_W1_MAX_INDX = 46
+
+# we are not using 'W2' as we're building
+# the histograms dynamically:
+HiCCUPS_W2_MAX_INDX = 10000
 
 
 def get_qvals(pvals):
@@ -21,7 +40,7 @@ def get_qvals(pvals):
     ----------
     pvals : array-like
         array of p-values to use for multiple hypothesis testing
-    
+
     Returns
     -------
     qvals : numpy.ndarray
@@ -29,19 +48,19 @@ def get_qvals(pvals):
 
     Notes
     -----
-    - BH-FDR reminder: given an array of N p-values, sort it in ascending order 
-    p[1]<p[2]<p[3]<...<p[N], and find a threshold p-value, p[j] for which 
-    p[j] < FDR*j/N, and p[j+1] is already p[j+1] >= FDR*(j+1)/N. Peaks 
+    - BH-FDR reminder: given an array of N p-values, sort it in ascending order
+    p[1]<p[2]<p[3]<...<p[N], and find a threshold p-value, p[j] for which
+    p[j] < FDR*j/N, and p[j+1] is already p[j+1] >= FDR*(j+1)/N. Peaks
     corresponding to p-values p[1]<p[2]<...p[j] are considered significant.
 
     - Mostly follows the statsmodels implementation:
     http://www.statsmodels.org/dev/_modules/statsmodels/stats/multitest.html
-    
-    - Using alpha=0.02 it is possible to achieve called dots similar to 
+
+    - Using alpha=0.02 it is possible to achieve called dots similar to
     pre-update status alpha is meant to be a q-value threshold: "qvals <= alpha"
-    
+
     '''
-    # NOTE: This is tested and is capable of reproducing 
+    # NOTE: This is tested and is capable of reproducing
     # `multiple_test_BH` results
     pvals = np.asarray(pvals)
     n_obs = pvals.size
@@ -64,8 +83,8 @@ def clust_2D_pixels(pixels_df,
                     verbose=True):
     '''
     Group significant pixels by proximity using Birch clustering. We use
-    "n_clusters=None", which implies no AgglomerativeClustering, and thus 
-    simply reporting "blobs" of pixels of radii <="threshold_cluster" along 
+    "n_clusters=None", which implies no AgglomerativeClustering, and thus
+    simply reporting "blobs" of pixels of radii <="threshold_cluster" along
     with corresponding blob-centroids as well.
 
     Parameters
@@ -89,7 +108,7 @@ def clust_2D_pixels(pixels_df,
         Name of the cluster of pixels size. "c_size" by default.
     verbose : bool
         Print verbose clustering summary report defaults is True.
-    
+
     Returns
     -------
     peak_tmp : pandas.DataFrame
@@ -108,8 +127,8 @@ def clust_2D_pixels(pixels_df,
     pixel_idxs = pixels_df.index
 
     # perform BIRCH clustering of pixels:
-    # "n_clusters=None" implies using BIRCH without AgglomerativeClustering, 
-    # thus simply reporting "blobs" of pixels of radius "threshold_cluster" 
+    # "n_clusters=None" implies using BIRCH without AgglomerativeClustering,
+    # thus simply reporting "blobs" of pixels of radius "threshold_cluster"
     # along with blob-centroids as well:
     brc = Birch(n_clusters=None,
                 threshold=threshold_cluster,
@@ -157,7 +176,7 @@ def clust_2D_pixels(pixels_df,
     centroids_n_labels_df[clust_label_name] = clustered_labels.astype(np.int)
     # add cluster sizes:
     centroids_n_labels_df[clust_size_name] = cluster_sizes.astype(np.int)
-    
+
 
     return centroids_n_labels_df
 
@@ -177,12 +196,12 @@ def diagonal_matrix_tiling(start, stop, bandwidth, edge=0, verbose=False):
     edge : int
         Small edge around each tile to be included in the yielded coordinates.
     bandwidth : int
-        Diagonal tiling is guaranteed to cover a diagonal are of width 
+        Diagonal tiling is guaranteed to cover a diagonal are of width
         'bandwidth'.
 
     Yields
     ------
-    Pairs of indices for every chunk use those indices [cstart:cstop) to fetch 
+    Pairs of indices for every chunk use those indices [cstart:cstop) to fetch
     chunks from the cooler-object:
     >>> clr.matrix()[cstart:cstop, cstart:cstop]
 
@@ -205,13 +224,13 @@ def diagonal_matrix_tiling(start, stop, bandwidth, edge=0, verbose=False):
     *                   *
     *                   *
     * * * * * * * * * * *
-    
-    These slices are supposed to cover up a diagonal band of size 'bandwidth'. 
+
+    These slices are supposed to cover up a diagonal band of size 'bandwidth'.
 
     """
     size = stop - start
     tiles = size // bandwidth + bool(size % bandwidth)
-    
+
     # matrix parameters before chunking:
     if verbose:
         print("matrix of size {}X{} to be split so that\n".format(size,size)+
@@ -220,11 +239,11 @@ def diagonal_matrix_tiling(start, stop, bandwidth, edge=0, verbose=False):
          "  a small 'edge' of size w={}, to allow for\n".format(edge)+
          "  meaningfull convolution around boundaries.\n"+
          "  Resulting number of tiles is {}".format(tiles-1)+
-         "  Non-edge case size of each tile is {}X{}".format(2*(bandwidth+edge), 
+         "  Non-edge case size of each tile is {}X{}".format(2*(bandwidth+edge),
                                                              2*(bandwidth+edge)))
 
     # actual number of tiles is tiles-1
-    # by doing range(1, tiles) we are making sure we are processing the 
+    # by doing range(1, tiles) we are making sure we are processing the
     # upper-left chunk only once:
     for t in range(1, tiles):
         # l = max(0,M*t-M)
@@ -237,7 +256,7 @@ def diagonal_matrix_tiling(start, stop, bandwidth, edge=0, verbose=False):
 
 def square_matrix_tiling(start, stop, step, edge, square=False, verbose=False):
     """
-    Generate a stream of tiling coordinates that guarantee to cover an entire 
+    Generate a stream of tiling coordinates that guarantee to cover an entire
     matrix. cis-signal only!
 
     Parameters
@@ -248,7 +267,7 @@ def square_matrix_tiling(start, stop, step, edge, square=False, verbose=False):
     stop : int
         End position of the matrix slice to be tiled (exclusive, bins, 0-based).
     step : int
-        Requested size of the tiles. Boundary tiles may or may not be of 
+        Requested size of the tiles. Boundary tiles may or may not be of
         'step', see 'square'.
     edge : int
         Small edge around each tile to be included in the yielded coordinates.
@@ -256,7 +275,7 @@ def square_matrix_tiling(start, stop, step, edge, square=False, verbose=False):
 
     Yields
     ------
-    Pairs of indices for every chunk use those indices [cstart:cstop) to fetch 
+    Pairs of indices for every chunk use those indices [cstart:cstop) to fetch
     chunks from the cooler-object:
     >>> clr.matrix()[cstart:cstop, cstart:cstop]
 
@@ -298,7 +317,7 @@ def square_matrix_tiling(start, stop, step, edge, square=False, verbose=False):
     *       *       *   *
     * * * * * * * * * * *
     WARNING: be carefull with extracting
-    expected in this case, as it is 
+    expected in this case, as it is
     not trivial at all !!!
 
     square = True
@@ -344,10 +363,10 @@ def square_matrix_tiling(start, stop, step, edge, square=False, verbose=False):
 
 def _convolve_and_count_nans(O_bal,E_bal,E_raw,N_bal,kernel):
     """
-    Dense versions of a bunch of matrices needed for convolution and 
-    calculation of number of NaNs in a vicinity of each pixel. And a kernel to 
+    Dense versions of a bunch of matrices needed for convolution and
+    calculation of number of NaNs in a vicinity of each pixel. And a kernel to
     be provided of course.
-    
+
     """
     # a matrix filled with the kernel-weighted sums
     # based on a balanced observed matrix:
@@ -367,24 +386,24 @@ def _convolve_and_count_nans(O_bal,E_bal,E_raw,N_bal,kernel):
     # pixel (kernel's nonzero footprint)
     # based on the NaN-matrix N_bal.
     # N_bal is shared NaNs between O_bal E_bal,
-    # is it redundant ? 
+    # is it redundant ?
     NN = convolve(N_bal.astype(np.int),
                   # we have to use kernel's
                   # nonzero footprint:
                   (kernel != 0).astype(np.int),
                   mode='constant',
-                  # there are only NaNs 
+                  # there are only NaNs
                   # beyond the boundary:
                   cval=1,
                   origin=0)
     ######################################
     # using cval=0 for actual data and
-    # cval=1 for NaNs matrix reduces 
+    # cval=1 for NaNs matrix reduces
     # "boundary issue" to the "number of
     # NaNs"-issue
     # ####################################
 
-    # now finally, E_raw*(KO/KE), as the 
+    # now finally, E_raw*(KO/KE), as the
     # locally-adjusted expected with raw counts as values:
     Ek_raw = np.multiply(E_raw, np.divide(KO, KE))
     # return locally adjusted expected and number of NaNs
@@ -434,10 +453,10 @@ def get_adjusted_expected_tile_some_nans(origin,
     Ek_raw[i,j] = E_raw[i,j]* ------------------ ,
                               KERNEL[i,j](E_bal)
     where E_raw[i,j] is:
-          1               1                 
+          1               1
     -------------- * -------------- * E_bal[i,j]
-    bal_weights[i]   bal_weights[j]             
-    
+    bal_weights[i]   bal_weights[j]
+
 
     Parameters
     ----------
@@ -466,7 +485,7 @@ def get_adjusted_expected_tile_some_nans(origin,
         into balanced observed for a slice of
         a matrix with the origin on the diagonal;
         and a tuple/list of a couple of 1D arrays
-        in case it is a slice with an arbitrary 
+        in case it is a slice with an arbitrary
         origin.
     kernels : dict of (str, numpy.ndarray)
         dictionary of kernels/masks to perform
@@ -480,9 +499,9 @@ def get_adjusted_expected_tile_some_nans(origin,
         Dictionay keys must contain names for
         each kernel.
         !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-        Beware!: kernels are flipped and 
+        Beware!: kernels are flipped and
         only then multiplied to matrix by
-        scipy.ndimage.convolve 
+        scipy.ndimage.convolve
         !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
     balance_factor: float
         Multiplicative Balancing factor:
@@ -498,7 +517,7 @@ def get_adjusted_expected_tile_some_nans(origin,
     verbose: bool
         Set to True to print some progress
         messages to stdout.
-    
+
     Returns
     -------
     peaks_df : pandas.DataFrame
@@ -507,7 +526,7 @@ def get_adjusted_expected_tile_some_nans(origin,
         for a given slice of input matrix. Multiple
         instences of such 'peaks_df' can be concatena-
         ted and deduplicated for the downstream analysis.
-        Reported columns: 
+        Reported columns:
         bin1_id - bin1_id index (row), adjusted to origin
         bin2_id - bin bin2_id index, adjusted to origin
         la_exp - locally adjusted expected (for each kernel)
@@ -562,13 +581,13 @@ def get_adjusted_expected_tile_some_nans(origin,
     N_bal = np.logical_or(np.isnan(O_bal),
                           np.isnan(E_bal))
     # fill in common nan-s with zeroes, preventing
-    # NaNs during convolution part of '_convolve_and_count_nans': 
+    # NaNs during convolution part of '_convolve_and_count_nans':
     O_bal[N_bal] = 0.0
     E_bal[N_bal] = 0.0
     # think about usinf copyto and where functions later:
     # https://stackoverflow.com/questions/6431973/how-to-copy-data-from-a-numpy-array-to-another
-    # 
-    # 
+    #
+    #
     # we are going to accumulate all the results
     # into a DataFrame, keeping NaNs, and other
     # unfiltered results (even the lower triangle for now):
@@ -594,8 +613,8 @@ def get_adjusted_expected_tile_some_nans(origin,
         #                                     E_raw,
         #                                     N_bal,
         #                                     kernel)
-        # Dense versions of a bunch of matrices needed for convolution and 
-        # calculation of number of NaNs in a vicinity of each pixel. And a kernel to 
+        # Dense versions of a bunch of matrices needed for convolution and
+        # calculation of number of NaNs in a vicinity of each pixel. And a kernel to
         # be provided of course.
         # a matrix filled with the kernel-weighted sums
         # based on a balanced observed matrix:
@@ -620,17 +639,17 @@ def get_adjusted_expected_tile_some_nans(origin,
                       # nonzero footprint:
                       (kernel != 0).astype(np.int),
                       mode='constant',
-                      # there are only NaNs 
+                      # there are only NaNs
                       # beyond the boundary:
                       cval=1,
                       origin=0)
         ######################################
         # using cval=0 for actual data and
-        # cval=1 for NaNs matrix reduces 
+        # cval=1 for NaNs matrix reduces
         # "boundary issue" to the "number of
         # NaNs"-issue
         # ####################################
-        # now finally, E_raw*(KO/KE), as the 
+        # now finally, E_raw*(KO/KE), as the
         # locally-adjusted expected with raw counts as values:
         Ek_raw = np.multiply(E_raw, np.divide(KO, KE))
 
@@ -685,10 +704,10 @@ def get_adjusted_expected_tile_some_nans(origin,
 
 def heatmap_tiles_generator_diag(clr, chroms, pad_size, tile_size, band_to_cover):
     """
-    A generator yielding heatmap tiles that are needed to cover the requested 
-    band_to_cover around diagonal. Each tile is "padded" with pad_size edge to 
+    A generator yielding heatmap tiles that are needed to cover the requested
+    band_to_cover around diagonal. Each tile is "padded" with pad_size edge to
     allow proper kernel-convolution of pixels close to boundary.
-    
+
     Parameters
     ----------
     clr : cooler
@@ -696,14 +715,14 @@ def heatmap_tiles_generator_diag(clr, chroms, pad_size, tile_size, band_to_cover
     chroms : iterable
         Iterable of chromosomes to process
     pad_size : int
-        Size of padding around each tile. Typically the outer size of the 
+        Size of padding around each tile. Typically the outer size of the
         kernel.
     tile_size : int
         Size of the heatmap tile.
     band_to_cover : int
-        Size of the diagonal band to be covered by the generated tiles. 
+        Size of the diagonal band to be covered by the generated tiles.
         Typically correspond to the max_loci_separation for called dots.
-        
+
     Returns
     -------
     tile : tuple
@@ -719,7 +738,7 @@ def heatmap_tiles_generator_diag(clr, chroms, pad_size, tile_size, band_to_cover
                                                  chr_stop,
                                                  tile_size,
                                                  pad_size):
-            # check if a given tile intersects with 
+            # check if a given tile intersects with
             # with the diagonal band of interest ...
             diag_from = tilej[0] - tilei[1]
             diag_to   = tilej[1] - tilei[0]
@@ -730,3 +749,753 @@ def heatmap_tiles_generator_diag(clr, chroms, pad_size, tile_size, band_to_cover
             # tiles from the lower triangle from calculations ...
             if (min(band_to,diag_to) - max(band_from,diag_from)) > 2*pad_size:
                 yield chrom, tilei, tilej
+
+
+def score_tile(tile_cij, clr, cis_exp, exp_v_name, bal_v_name, kernels,
+               nans_tolerated, band_to_cover, balance_factor, verbose):
+    """
+    The main working function that given a tile of a heatmap, applies kernels to
+    perform convolution to calculate locally-adjusted expected and then
+    calculates a p-value for every meaningfull pixel against these l.a. expected
+    values.
+
+    Parameters
+    ----------
+    tile_cij : tuple
+        Tuple of 3: chromosome name, tile span row-wise, tile span column-wise:
+        (chrom, tile_i, tile_j), where tile_i = (start_i, end_i), and
+        tile_j = (start_j, end_j).
+    clr : cooler
+        Cooler object to use to extract Hi-C heatmap data.
+    cis_exp : pandas.DataFrame
+        DataFrame with 1 dimensional expected, indexed with 'chrom' and 'diag'.
+    exp_v_name : str
+        Name of a value column in expected DataFrame
+    bal_v_name : str
+        Name of a value column with balancing weights in a cooler.bins()
+        DataFrame. Typically 'weight'.
+    kernels : dict
+        A dictionary with keys being kernels names and values being ndarrays
+        representing those kernels.
+    nans_tolerated : int
+        Number of NaNs tolerated in a footprint of every kernel.
+    band_to_cover : int
+        Results would be stored only for pixels connecting loci closer than
+        'band_to_cover'.
+    balance_factor : float
+        Balancing factor to turn sum of balanced matrix back approximately
+        to the number of pairs (used for dynamic-donut criteria mostly).
+        use None value to disable dynamic-donut criteria calculation.
+    verbose : bool
+        Enable verbose output.
+
+    Returns
+    -------
+    res_df : pandas.DataFrame
+        results: annotated pixels with calculated locally adjusted expected
+        for every kernels, observed, precalculated pvalues, number of NaNs in
+        footprint of every kernels, all of that in a form of an annotated
+        pixels DataFrame for eligible pixels of a given tile.
+
+    """
+    # unpack tile's coordinates
+    chrom, tilei, tilej = tile_cij
+    origin = (tilei[0], tilej[0])
+
+    # we have to do it for every tile, because
+    # chrom is not known apriori (maybe move outside):
+    lazy_exp = LazyToeplitz(cis_exp.loc[chrom][exp_v_name].values)
+
+    # RAW observed matrix slice:
+    observed = clr.matrix(balance=False)[slice(*tilei), slice(*tilej)]
+    # expected as a rectangular tile :
+    expected = lazy_exp[slice(*tilei), slice(*tilej)]
+    # slice of balance_weight for row-span and column-span :
+    bal_weight_i = clr.bins()[slice(*tilei)][bal_v_name].values
+    bal_weight_j = clr.bins()[slice(*tilej)][bal_v_name].values
+
+    # do the convolutions
+    result = get_adjusted_expected_tile_some_nans(
+        origin=origin,
+        observed=observed,
+        expected=expected,
+        bal_weights=(bal_weight_i,bal_weight_j),
+        kernels=kernels,
+        balance_factor=balance_factor,
+        verbose=verbose)
+
+    # Post-processing filters
+    # (1) exclude pixels that connect loci further than 'band_to_cover' apart:
+    is_inside_band = (result["bin1_id"] > (result["bin2_id"]-band_to_cover))
+
+    # (2) identify pixels that pass number of NaNs compliance test for ALL kernels:
+    does_comply_nans = np.all(
+        result[["la_exp."+k+".nnans" for k in kernels]] < nans_tolerated,
+        axis=1)
+    # so, selecting inside band and nNaNs compliant results:
+    # ( drop dropping index maybe ??? ) ...
+    res_df = result[is_inside_band & does_comply_nans].reset_index(drop=True)
+    ########################################################################
+    # consider retiring Poisson testing from here, in case we
+    # stick with l-chunking or opposite - add histogramming business here(!)
+    ########################################################################
+    # do Poisson tests:
+    get_pval = lambda la_exp : 1.0 - poisson.cdf(res_df["obs.raw"], la_exp)
+    for k in kernels:
+        res_df["la_exp."+k+".pval"] = get_pval( res_df["la_exp."+k+".value"] )
+
+    # annotate and return
+    return cooler.annotate(res_df.reset_index(drop=True), clr.bins()[:])
+
+
+def histogram_scored_pixels(scored_df, kernels, ledges, verbose):
+    """
+    An attempt to implement HiCCUPS-like lambda-chunking
+    statistical procedure.
+    This function aims at building up a histogram of locally
+    adjusted expected scores for groups of characterized
+    pixels.
+    Such histograms are later used to compute FDR thresholds
+    for different "classes" of hypothesis (classified by their
+    l.a. expected scores).
+
+    Parameters
+    ----------
+    scored_df : pd.DataFrame
+        A table with the scoring information for a group of pixels.
+    kernels : dict
+        A dictionary with keys being kernels names and values being ndarrays
+        representing those kernels.
+    ledges : ndarray
+        An ndarray with bin lambda-edges for groupping loc. adj. expecteds,
+        i.e., classifying statistical hypothesis into lambda-classes.
+        Left-most bin (-inf, 1], and right-most one (value,+inf].
+    verbose : bool
+        Enable verbose output.
+
+    Returns
+    -------
+    hists : dict of pandas.DataFrame
+        A dictionary of pandas.DataFrame with lambda/observed histogram for
+        every kernel-type.
+
+
+    Notes
+    -----
+    This is just an attempt to implement HiCCUPS-like lambda-chunking.
+    So we'll be returning histograms corresponding to the chunks of
+    scored pixels.
+    Consider modifying/accumulation a globally defined hists object,
+    or probably making something like a Feature2D class later on
+    where hists would be a class feature and histogramming_step would be
+    a method.
+
+
+    """
+
+    # lambda-chunking implies different 'pval' calculation
+    # procedure with a single Poisson expected for all the
+    # hypothesis in a same "class", i.e. with the l.a. expecteds
+    # from the same histogram bin.
+
+
+    ########################
+    # implementation ideas:
+    ########################
+    # same observations/hypothesis needs to be classified according
+    # to different l.a. expecteds (i.e. for different kernel-types),
+    # which can be done with a pandas groupby, something like that:
+    # https://stackoverflow.com/questions/21441259
+    #
+    # after that we could iterate over groups and do np.bincout on
+    # the "observed.raw" column (assuming it's of integer type) ...
+
+    hists = {}
+    for k in kernels:
+        # verbose:
+        if verbose:
+            print("Building a histogram for kernel-type {}".format(k))
+        #  we would need to generate a bunch of these histograms for all of the
+        # kernel types:
+        # needs to be lambda-binned             : scored_df["la_exp."+k+".value"]
+        # needs to be histogrammed in every bin : scored_df["obs.raw"]
+        #
+        # lambda-bin index for kernel-type "k":
+        lbins = pd.cut(scored_df["la_exp."+k+".value"],ledges)
+        # now for each lambda-bin construct a histogramm of "obs.raw":
+        obs_hist = {}
+        for lbin, grp_df in scored_df.groupby(lbins):
+            # check if obs.raw is integer of spome kind (temporary):
+            assert np.issubdtype(grp_df["obs.raw"].dtype, np.integer)
+            # perform bincounting ...
+            obs_hist[lbin] = pd.Series(np.bincount(grp_df["obs.raw"]))
+            # ACHTUNG! assigning directly to empty DF leads to data loss!
+            # turn ndarray in Series for ease of handling, i.e.
+            # assign a bunch of columns of different sizes to DataFrame.
+            #
+            # Consider updating global "hists" later on, or implementing a
+            # special class for that. Mind that Python multiprocessing
+            # "threads" are processes and thus cannot share/modify a shared
+            # memory location - deal with it, maybe dask-something ?!
+            #
+            # turned out that storing W1xW2 for every "thread" requires a ton
+            # of memory - that's why different sizes... @nvictus ?
+        # store W1x(<=W2) hist for every kernel-type:
+        hists[k] = pd.DataFrame(obs_hist).fillna(0).astype(np.integer)
+    # return a dict of DataFrames with a bunch of histograms:
+    return hists
+
+
+def extract_scored_pixels(scored_df, kernels, thresholds, ledges, verbose):
+    """
+    An attempt to implement HiCCUPS-like lambda-chunking
+    statistical procedure.
+    Use FDR thresholds for different "classes" of hypothesis
+    (classified by their l.a. expected scores), in order to
+    extract pixels that stand out.
+
+    Parameters
+    ----------
+    scored_df : pd.DataFrame
+        A table with the scoring information for a group of pixels.
+    kernels : dict
+        A dictionary with keys being kernel names and values being ndarrays
+        representing those kernels.
+    thresholds : dict
+        A dictionary with keys being kernel names and values pandas.Series
+        indexed with Intervals defined by 'ledges' boundaries and storing FDR
+        thresholds for observed values.
+    ledges : ndarray
+        An ndarray with bin lambda-edges for groupping loc. adj. expecteds,
+        i.e., classifying statistical hypothesis into lambda-classes.
+        Left-most bin (-inf, 1], and right-most one (value,+inf].
+    verbose : bool
+        Enable verbose output.
+
+    Returns
+    -------
+    scored_df_slice : pandas.DataFrame
+        Filtered DataFrame of pixels extracted applying FDR thresholds.
+
+    Notes
+    -----
+    This is just an attempt to implement HiCCUPS-like lambda-chunking.
+
+    """
+    comply_fdr_list = np.ones(len(scored_df), dtype=np.bool)
+
+    for k in kernels:
+        # using special features of IntervalIndex:
+        # http://pandas.pydata.org/pandas-docs/stable/advanced.html#intervalindex
+        # i.e. IntervalIndex can be .loc-ed with values that would be
+        # corresponded with their Intervals (!!!):
+        comply_fdr_k = (scored_df["obs.raw"].values > \
+                        thresholds[k].loc[scored_df["la_exp."+k+".value"]].values)
+        ## attempting to extract q-values using l-chunks and IntervalIndex:
+        ## we'll do it in an ugly but workign fashion, by simply
+        ## iteration over pairs of obs, la_exp and extracting needed qvals
+        ## one after another ...
+        #scored_df["la_exp."+k+".qval"] = \
+        #        [ qvalues[k].loc[o,e] for o,e \
+        #            in scored_df[["obs.raw","la_exp."+k+".value"]].itertuples(index=False) ]
+        ##
+        ## accumulate comply_fdr_k into comply_fdr_list
+        # using np.logical_and:
+        comply_fdr_list = np.logical_and(comply_fdr_list, comply_fdr_k)
+    # return a slice of 'scored_df' that complies FDR thresholds:
+    return scored_df[comply_fdr_list]
+
+
+def scoring_step(clr, expected, expected_name, tiles, kernels,
+                 max_nans_tolerated, loci_separation_bins, output_path,
+                 nproc, verbose):
+    if verbose:
+        print("Preparing to convolve {} tiles:".format(len(tiles)))
+
+    # add very_verbose to supress output from convolution of every tile
+    very_verbose = False
+    job = partial(
+        score_tile,
+        clr=clr,
+        cis_exp=expected,
+        exp_v_name=expected_name,
+        bal_v_name='weight',
+        kernels=kernels,
+        nans_tolerated=max_nans_tolerated,
+        band_to_cover=loci_separation_bins,
+        verbose=very_verbose)
+
+    if nproc > 1:
+        pool = mp.Pool(nproc)
+        map_ = pool.imap
+        map_kwargs = dict(chunksize=int(np.ceil(len(tiles)/nproc)))
+        if verbose:
+            print("creating a Pool of {} workers to tackle {} tiles".format(
+                    nproc, len(tiles)))
+    else:
+        map_ = map
+        if verbose:
+            print("fallback to serial implementation.")
+        map_kwargs = {}
+    try:
+        # consider using
+        # https://github.com/mirnylab/cooler/blob/9e72ee202b0ac6f9d93fd2444d6f94c524962769/cooler/tools.py#L59
+        # here:
+        chunks = map_(job, tiles, **map_kwargs)
+        append = False
+        for chunk in chunks:
+            chunk.to_hdf(output_path,
+                         key='results',
+                         format='table',
+                         append=append)
+            append = True
+    finally:
+        if nproc > 1:
+            pool.close()
+
+
+def scoring_and_histogramming_step(clr, expected, expected_name, tiles, kernels,
+                                   ledges, max_nans_tolerated, loci_separation_bins,
+                                   output_path, nproc, verbose):
+    """
+    This is a derivative of the 'scoring_step'
+    which is supposed to implement the 1st of the
+    lambda-chunking procedure - histogramming.
+
+    Basically we are piping scoring operation
+    together with histogramming into a single
+    pipeline of per-chunk operations/transforms.
+    """
+    if verbose:
+        print("Preparing to convolve {} tiles:".format(len(tiles)))
+
+    # add very_verbose to supress output from convolution of every tile
+    very_verbose = False
+
+    # to score per tile:
+    to_score = partial(
+        score_tile,
+        clr=clr,
+        cis_exp=expected,
+        exp_v_name=expected_name,
+        bal_v_name='weight',
+        kernels=kernels,
+        nans_tolerated=max_nans_tolerated,
+        band_to_cover=loci_separation_bins,
+        # do not calculate dynamic-donut criteria
+        # for now.
+        balance_factor=None,
+        verbose=very_verbose)
+
+    # to hist per scored chunk:
+    to_hist = partial(
+        histogram_scored_pixels,
+        kernels=kernels,
+        ledges=ledges,
+        verbose=very_verbose)
+
+    # composing/piping scoring and histogramming
+    # together :
+    job = lambda tile : to_hist(to_score(tile))
+
+    # copy paste from @nvictus modified 'scoring_step':
+    if nproc > 1:
+        pool = mp.Pool(nproc)
+        map_ = pool.imap
+        map_kwargs = dict(chunksize=int(np.ceil(len(tiles)/nproc)))
+        if verbose:
+            print("creating a Pool of {} workers to tackle {} tiles".format(
+                    nproc, len(tiles)))
+    else:
+        map_ = map
+        if verbose:
+            print("fallback to serial implementation.")
+        map_kwargs = {}
+    try:
+        # consider using
+        # https://github.com/mirnylab/cooler/blob/9e72ee202b0ac6f9d93fd2444d6f94c524962769/cooler/tools.py#L59
+        # here:
+        hchunks = map_(job, tiles, **map_kwargs)
+        # hchunks TO BE ACCUMULATED
+        # hopefully 'hchunks' would stay in memory
+        # until we would get a chance to accumulate them:
+    finally:
+        if nproc > 1:
+            pool.close()
+    #
+    # now we need to combine/sum all of the histograms
+    # for different kernels:
+    #
+    # assuming we know "kernels"
+    # this is very ugly, but ok
+    # for the draft lambda-chunking
+    # lambda version of lambda-chunking:
+    def _sum_hists(hx,hy):
+        # perform a DataFrame summation
+        # for every value of the dictionary:
+        hxy = {}
+        for k in kernels:
+            hxy[k] = hx[k].add(hy[k],fill_value=0).astype(np.integer)
+        # returning the sum:
+        return hxy
+
+    # ######################################################
+    # this approach is tested and at the very least
+    # number of pixels in a dump list matches
+    # with the .sum().sum() of the histogram
+    # both for 10kb and 5kb,
+    # thus we should consider this as a reference
+    # implementation, albeit not a very efficient one ...
+    # ######################################################
+    final_hist = reduce(_sum_hists, hchunks)
+    # we have to make sure there is nothing in the
+    # top bin, i.e., there are no l.a. expecteds > base^(len(ledges)-1)
+    for k in kernels:
+        last_la_exp_bin = final_hist[k].columns[-1]
+        last_la_exp_vals = final_hist[k].iloc[:,-1]
+        # checking the top bin:
+        assert (last_la_exp_vals.sum()==0), \
+                "There are la_exp.{}.value in {}, please check the histogram" \
+                                                    .format(k,last_la_exp_bin)
+        # drop that last column/bin (last_edge, +inf]:
+        final_hist[k] = final_hist[k].drop(columns=last_la_exp_bin)
+        # consider dropping all of the columns that have zero .sum()
+    # returning filtered histogram
+    return final_hist
+
+
+def scoring_and_extraction_step(clr, expected, expected_name, tiles, kernels,
+                               ledges, thresholds, max_nans_tolerated,
+                               balance_factor, loci_separation_bins, output_path,
+                               nproc, verbose):
+    """
+    This is a derivative of the 'scoring_step'
+    which is supposed to implement the 2nd of the
+    lambda-chunking procedure - extracting pixels
+    that are FDR compliant.
+
+    Basically we are piping scoring operation
+    together with extraction into a single
+    pipeline of per-chunk operations/transforms.
+    """
+    if verbose:
+        print("Preparing to convolve {} tiles:".format(len(tiles)))
+
+    # add very_verbose to supress output from convolution of every tile
+    very_verbose = False
+
+    # to score per tile:
+    to_score = partial(
+        score_tile,
+        clr=clr,
+        cis_exp=expected,
+        exp_v_name=expected_name,
+        bal_v_name='weight',
+        kernels=kernels,
+        nans_tolerated=max_nans_tolerated,
+        band_to_cover=loci_separation_bins,
+        balance_factor=balance_factor,
+        verbose=very_verbose)
+
+    # to hist per scored chunk:
+    to_extract = partial(
+        extract_scored_pixels,
+        kernels=kernels,
+        thresholds=thresholds,
+        ledges=ledges,
+        verbose=very_verbose)
+
+    # composing/piping scoring and histogramming
+    # together :
+    job = lambda tile : to_extract(to_score(tile))
+
+    # copy paste from @nvictus modified 'scoring_step':
+    if nproc > 1:
+        pool = mp.Pool(nproc)
+        map_ = pool.imap
+        map_kwargs = dict(chunksize=int(np.ceil(len(tiles)/nproc)))
+        if verbose:
+            print("creating a Pool of {} workers to tackle {} tiles".format(
+                    nproc, len(tiles)))
+    else:
+        map_ = map
+        if verbose:
+            print("fallback to serial implementation.")
+        map_kwargs = {}
+    try:
+        # consider using
+        # https://github.com/mirnylab/cooler/blob/9e72ee202b0ac6f9d93fd2444d6f94c524962769/cooler/tools.py#L59
+        # here:
+        filtered_pix_chunks = map_(job, tiles, **map_kwargs)
+        significant_pixels = pd.concat(filtered_pix_chunks,ignore_index=True)
+        if output_path is not None:
+            significant_pixels.to_csv(output_path,
+                                      sep='\t',
+                                      header=True,
+                                      index=False,
+                                      compression=None)
+    finally:
+        if nproc > 1:
+            pool.close()
+    # # concat and store the results if needed:
+    # significant_pixels = pd.concat(filtered_pix_chunks)
+    return significant_pixels \
+                .sort_values(by=["chrom1","chrom2","start1","start2"]) \
+                .reset_index(drop=True)
+
+
+
+def clustering_step_local(scores_df, expected_chroms,
+                          dots_clustering_radius, verbose):
+    """
+    This is a new "clustering" step updated for the pixels
+    processed by lambda-chunking multiple hypothesis testing.
+
+    This method assumes that 'scores_df' is a DataFrame with
+    all of the pixels that needs to be clustered, thus there is
+    no additional 'comply_fdr' column and selection of compliant
+    pixels.
+
+    This step is a clustering-only (using Birch from scikit).
+
+    Parameters
+    ----------
+    scores_df : pandas.DataFrame
+        DataFrame that stores filtered pixels that are ready to be
+        clustered, no more 'comply_fdr' column dependency.
+    expected_chroms : iterable
+        An iterable of chromosomes to be clustered.
+    dots_clustering_radius : int
+        Birch-clustering threshold.
+    verbose : bool
+        Enable verbose output.
+
+    Returns
+    -------
+    centroids : pandas.DataFrame
+        Pixels from 'scores_df' annotated with clustering
+        information.
+
+    Notes
+    -----
+    'dots_clustering_radius' in Birch clustering algorithm
+    corresponds to a double the clustering radius in the
+    "greedy"-clustering used in HiCCUPS (to be tested).
+
+    """
+
+    # using different bin12_id_names since all
+    # pixels are annotated at this point.
+    pixel_clust_list = []
+    for chrom  in expected_chroms:
+        # probably generate one big DataFrame with clustering
+        # information only and then just merge it with the
+        # existing 'scores_df'-DataFrame.
+        # should we use groupby instead of 'scores_df['chrom12']==chrom' ?!
+        # to be tested ...
+        df = scores_df[((scores_df['chrom1'].astype(str)==str(chrom)) &
+                        (scores_df['chrom2'].astype(str)==str(chrom)))]
+
+        pixel_clust = clust_2D_pixels(
+            df,
+            threshold_cluster=dots_clustering_radius,
+            bin1_id_name='start1',
+            bin2_id_name='start2',
+            verbose=verbose)
+        pixel_clust_list.append(pixel_clust)
+    if verbose:
+        print("Clustering is over!")
+    # concatenate clustering results ...
+    # indexing information persists here ...
+    pixel_clust_df = pd.concat(pixel_clust_list, ignore_index=False)
+
+    # now merge pixel_clust_df and scores_df DataFrame ...
+    # # and merge (index-wise) with the main DataFrame:
+    df = pd.merge(
+        scores_df,
+        pixel_clust_df,
+        how='left',
+        left_index=True,
+        right_index=True)
+
+    # report only centroids with highest Observed:
+    chrom_clust_group = df.groupby(["chrom1", "chrom2", "c_label"])
+    centroids = df.loc[chrom_clust_group["obs.raw"].idxmax()]
+    return centroids
+
+
+def clustering_step(scores_file, expected_chroms, ktypes, fdr,
+                    dots_clustering_radius, verbose):
+    """
+    This is an old "clustering" step, before lambda-chunking
+    was implemented.
+    This step actually includes both multiple hypothesis
+    testing (its simple genome-wide version of BH-FDR) and
+    a clustering step itself (using Birch from scikit).
+    This method also assumes 'scores_file' to be an external
+    hdf file, and it would try to read the entire file in
+    memory.
+    """
+    res_df = pd.read_hdf(scores_file, 'results')
+
+    # do Benjamin-Hochberg FDR multiple hypothesis tests
+    # genome-wide:
+    for k in ktypes:
+        res_df["la_exp."+k+".qval"] = get_qvals( res_df["la_exp."+k+".pval"] )
+
+    # combine results of all tests:
+    res_df['comply_fdr'] = np.all(
+        res_df[["la_exp."+k+".qval" for k in ktypes]] <= fdr,
+        axis=1)
+
+    # print a message for timing:
+    if verbose:
+        print("Genome-wide multiple hypothesis testing is done.")
+
+    # using different bin12_id_names since all
+    # pixels are annotated at this point.
+    pixel_clust_list = []
+    for chrom  in expected_chroms:
+        # probably generate one big DataFrame with clustering
+        # information only and then just merge it with the
+        # existing 'res_df'-DataFrame.
+        # should we use groupby instead of 'res_df['chrom12']==chrom' ?!
+        # to be tested ...
+        df = res_df[(res_df['comply_fdr'] &
+                    (res_df['chrom1']==chrom) &
+                    (res_df['chrom2']==chrom))]
+
+        pixel_clust = clust_2D_pixels(
+            df,
+            threshold_cluster=dots_clustering_radius,
+            bin1_id_name='start1',
+            bin2_id_name='start2',
+            verbose=verbose)
+        pixel_clust_list.append(pixel_clust)
+    if verbose:
+        print("Clustering is over!")
+    # concatenate clustering results ...
+    # indexing information persists here ...
+    pixel_clust_df = pd.concat(pixel_clust_list, ignore_index=False)
+
+    # now merge pixel_clust_df and res_df DataFrame ...
+    # # and merge (index-wise) with the main DataFrame:
+    df = pd.merge(
+        res_df[res_df['comply_fdr']],
+        pixel_clust_df,
+        how='left',
+        left_index=True,
+        right_index=True)
+
+    # report only centroids with highest Observed:
+    chrom_clust_group = df.groupby(["chrom1", "chrom2", "c_label"])
+    centroids = df.loc[chrom_clust_group["obs.raw"].idxmax()]
+    return centroids
+
+
+def thresholding_step(centroids):
+    # (2)
+    # filter by FDR, enrichment etc:
+    enrichment_factor_1 = 1.5
+    enrichment_factor_2 = 1.75
+    enrichment_factor_3 = 2.0
+    FDR_orphan_threshold = 0.02
+    ######################################################################
+    # # Temporarily remove orphans filtering, until q-vals are calculated:
+    ######################################################################
+    enrichment_fdr_comply = (
+        (centroids["obs.raw"] > enrichment_factor_2 * centroids["la_exp.lowleft.value"]) &
+        (centroids["obs.raw"] > enrichment_factor_2 * centroids["la_exp.donut.value"]) &
+        (centroids["obs.raw"] > enrichment_factor_1 * centroids["la_exp.vertical.value"]) &
+        (centroids["obs.raw"] > enrichment_factor_1 * centroids["la_exp.horizontal.value"]) &
+        ( (centroids["obs.raw"] > enrichment_factor_3 * centroids["la_exp.lowleft.value"])
+            | (centroids["obs.raw"] > enrichment_factor_3 * centroids["la_exp.donut.value"]) ) &
+        ( (centroids["c_size"] > 1)
+           | ((centroids["la_exp.lowleft.qval"]
+               + centroids["la_exp.donut.qval"]
+               + centroids["la_exp.vertical.qval"]
+               + centroids["la_exp.horizontal.qval"]) <= FDR_orphan_threshold)
+        )
+    )
+    # #
+    # enrichment_fdr_comply = (
+    #     (centroids["obs.raw"] > enrichment_factor_2 * centroids["la_exp.lowleft.value"]) &
+    #     (centroids["obs.raw"] > enrichment_factor_2 * centroids["la_exp.donut.value"]) &
+    #     (centroids["obs.raw"] > enrichment_factor_1 * centroids["la_exp.vertical.value"]) &
+    #     (centroids["obs.raw"] > enrichment_factor_1 * centroids["la_exp.horizontal.value"]) &
+    #     ( (centroids["obs.raw"] > enrichment_factor_3 * centroids["la_exp.lowleft.value"])
+    #         | (centroids["obs.raw"] > enrichment_factor_3 * centroids["la_exp.donut.value"]) )
+    # )
+    # use "enrichment_fdr_comply" to filter out
+    # non-satisfying pixels:
+    out = centroids[enrichment_fdr_comply]
+
+    # # # columns up for grabs, take whatever you need:
+    # chrom1
+    # start1
+    # end1
+    # weight1
+    # chrom2
+    # start2
+    # end2
+    # weight2
+    # la_exp.donut.value
+    # la_exp.donut.nnans
+    # la_exp.vertical.value
+    # la_exp.vertical.nnans
+    # la_exp.horizontal.value
+    # la_exp.horizontal.nnans
+    # la_exp.lowleft.value
+    # la_exp.lowleft.nnans
+    # la_exp.upright.value
+    # la_exp.upright.nnans
+    # exp.raw
+    # obs.raw
+    # la_exp.donut.pval
+    # la_exp.vertical.pval
+    # la_exp.horizontal.pval
+    # la_exp.lowleft.pval
+    # la_exp.upright.pval
+    # la_exp.donut.qval
+    # la_exp.vertical.qval
+    # la_exp.horizontal.qval
+    # la_exp.lowleft.qval
+    # la_exp.upright.qval
+    # comply_fdr
+    # cstart1
+    # cstart2
+    # c_label
+    # c_size
+
+    # ...
+    # to be added to the list of output columns:
+    # "factor_balance."+"lowleft"+".KerObs"
+    # ...
+
+    # tentaive output columns list:
+    columns_for_output = [
+        'chrom1',
+        'start1',
+        'end1',
+        'chrom2',
+        'start2',
+        'end2',
+        'cstart1',
+        'cstart2',
+        'c_label',
+        'c_size',
+        'obs.raw',
+        'exp.raw',
+        'la_exp.donut.value',
+        'la_exp.vertical.value',
+        'la_exp.horizontal.value',
+        'la_exp.lowleft.value',
+        "factor_balance.lowleft.KerObs",
+        # 'la_exp.upright.value',
+        # 'la_exp.upright.qval',
+        'la_exp.donut.qval',
+        'la_exp.vertical.qval',
+        'la_exp.horizontal.qval',
+        'la_exp.lowleft.qval'
+    ]
+    return out[columns_for_output]
